project/requirements.txt
asgiref
Django
django-debug-toolbar
django-jalali
easy-thumbnails
jalali_core
jdatetime
pillow
schedule
sqlparse
tzdata
psycopg2-binary
weasyprint